requirements.txt
streamlit 
websockets 
python-dotenv
fastapi 
uvicorn
python-multipart
httpx